break out importer view
