+++ Tests/WolvenKit.Utility/appsettings.json
@@ -1,5 +1,5 @@
 {
-  "GameDirectory": "Z:\\GOG Galaxy\\Games\\Cyberpunk 2077",
+  "GameDirectory": "D:\\gog\\Cyberpunk 2077",
   "WriteToFile": true,
   "LIMIT": 5000,
   "KEEP": true
