+++ WolvenKit/GenericHost.cs
@@ -8,6 +8,7 @@ using Splat.Microsoft.Extensions.DependencyInjection;
 using WolvenKit.App;
 using WolvenKit.App.Services;
 using WolvenKit.App.ViewModels.Dialogs;
+using WolvenKit.App.ViewModels.Importers;
 using WolvenKit.Common;
 using WolvenKit.Common.Interfaces;
 using WolvenKit.Common.Services;
@@ -30,6 +31,7 @@ using WolvenKit.Views.Dialogs;
 using WolvenKit.Views.Dialogs.Windows;
 using WolvenKit.Views.HomePage;
 using WolvenKit.Views.HomePage.Pages;
+using WolvenKit.Views.Importers;
 using WolvenKit.Views.Shell;
 using WolvenKit.Views.Tools;
 using WolvenKit.Views.Wizards;
@@ -143,6 +145,9 @@ namespace WolvenKit
                     services.AddTransient<ProjectWizardViewModel>();
                     services.AddTransient<IViewFor<ProjectWizardViewModel>, ProjectWizardView>();
 
+                    services.AddTransient<TextureImportViewModel>();
+                    services.AddTransient<IViewFor<TextureImportViewModel>, TextureImportView>();
+
                     #endregion
 
                     #region documents
@@ -175,8 +180,8 @@ namespace WolvenKit
                     //services.AddTransient<VisualEditorViewModel>();
                     //services.AddTransient<IViewFor<VisualEditorViewModel>, VisualEditorView>();
 
-                    services.AddSingleton<ImportExportViewModel>();
-                    services.AddTransient<IViewFor<ImportExportViewModel>, ImportExportView>();
+                    //services.AddSingleton<ImportExportViewModel>();
+                    //services.AddTransient<IViewFor<ImportExportViewModel>, ImportExportView>();
 
                     #endregion
